requirements.txt
asgiref==3.3.0
Django==3.1.2
django-jsonfield==1.4.0
django-log-request-id==1.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
